requirements.txt
python-dotenv==0.10.3
python-telegram-bot==11.1.0
requests==2.22.0
geopy==1.20.0
cyrtranslit==0.4
redis==3.3.8
